+++ Server/GameServer/game/ms_coordinate.h
@@ -2,8 +2,8 @@
 #include <cstdint>
 
 struct MsCoordinate {
-  double x;
-  double y;
+  float x;
+  float y;
   int16_t grid_x;
   int16_t grid_y;
   int16_t grid_size = 1;
